+++ tests/testthat/test-messages.R
@@ -1,0 +1,16 @@
+
+context("messages")
+
+test_that("messages in callr::r do not crash session", {
+  ret <- r(function() { cliapp::cli_text("fooobar"); 1 + 1 })
+  expect_identical(ret, 2)
+})
+
+test_that("messages in callr::r_bg do not crash session", {
+  rx <- r_bg(function() { cliapp::cli_text("fooobar"); 1 + 1 })
+  rx$wait(2000)
+  rx$kill()
+  expect_equal(rx$get_exit_status(), 0)
+
+  expect_equal(rx$get_result(), 2)
+})