+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # dev
 
+* `r()`, `r_bg()`, etc. now handle messages from the cliapp package
+  properly. They used to make the R session exit.
+
 # callr 3.1.0
 
 * New `rscript()` function and `rscript_process` class to execute